% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backtest_threshold}
\alias{backtest_threshold}
\title{Backtest Threshold Function}
\usage{
backtest_threshold(returns, indicator, threshold, sell_below)
}
\arguments{
\item{returns}{A numeric vector of returns.}

\item{indicator}{A numeric vector used as an indicator.}

\item{threshold}{A numeric threshold.}

\item{sell_below}{Logical, indicating sell decision.}
}
\value{
A single numeric value corresponding to the adjusted cumulative return.
}
\description{
This function computes cumulative returns.
}
