% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse.R
\name{coarse}
\alias{coarse}
\title{Coarse filtering}
\usage{
coarse(
  min_mean_mon_price = 1,
  min_mean_mon_volume = 10000,
  min_last_mon_mcap = 1e+09,
  dollar_vol_n = NULL,
  ...
)
}
\arguments{
\item{min_mean_mon_price}{Minimum mean monthly price (default 1)}

\item{min_mean_mon_volume}{Minimum mean monthly volume (default 10000)}

\item{min_last_mon_mcap}{Minimum last month market cap (default 1e9)}

\item{dollar_vol_n}{If not NULL, select top n by dollar volume each month}

\item{...}{Arguments passed to \code{qc_daily_parquet}}
}
\value{
Data frame of daily prices
}
\description{
Coarse equity filtering.
}
