% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_daily.R
\name{crypto}
\alias{crypto}
\title{Import crypto data}
\usage{
crypto(
  path_market_cap,
  path_prices,
  first_date = as.Date("2018-08-05"),
  snapshot_hour = 0,
  n = 10,
  min_constituents = 10
)
}
\arguments{
\item{path_market_cap}{Character. Path to market cap data.}

\item{path_prices}{Character. Path to prices data.}

\item{first_date}{Date. First date to look at.}

\item{snapshot_hour}{Numeric. Hours to keep from hour data.}

\item{n}{Integer Number of coins to keep from market cap}

\item{min_constituents}{Integer. Minimum number of constituents.}
}
\value{
A cleaned and processed data.table with prices and universe data.
}
\description{
This function processes hour crypto makret data and marcet cap data
}
