% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dukascopy_quotes.R
\name{dukascopy_quotes}
\alias{dukascopy_quotes}
\title{Process and Clean dukascopy Tick Quotes data}
\usage{
dukascopy_quotes(dir_path, market = c("ususd"), symbols = NULL, clean = FALSE)
}
\arguments{
\item{dir_path}{Character. Path to the directory containing the raw price data.}

\item{market}{Character. Market of the quotes. Default is "ususd".}

\item{symbols}{Character. Symbols to process. Default is NULL.}

\item{clean}{Logical. Clean quotes using highfrequency package. Default is FALSE.}
}
\value{
A cleaned and processed data.table with price and return information.
}
\description{
This function processes and cleans raw tick quotes data from Dukascopy.
}
