% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finam_ohlcv.R
\name{finam_ohlcv}
\alias{finam_ohlcv}
\title{Process and Clean Dukascopy Tick Quotes data}
\usage{
finam_ohlcv(dir_path, symbols, clean = FALSE)
}
\arguments{
\item{dir_path}{A character string with the path to the directory containing the raw tick quotes data.}

\item{symbols}{A character string with the symbol to process.
If NULL, all symbols in the directory will be processed.}

\item{clean}{A logical value indicating whether to clean the data using the
highfrequency package. Default is FALSE.}
}
\value{
A cleaned and processed data.table with price and return information.
}
\description{
This function processes and cleans raw tick quotes data from Dukascopy.
}
