% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finutils-package.R
\docType{package}
\name{finutils-package}
\alias{finutils}
\alias{finutils-package}
\title{finutils: Utils functions for quant finance}
\description{
Utils functions I often use in my quant finance analysis.
}
\keyword{internal}
