% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_stats.R
\name{portfolio_stats}
\alias{portfolio_stats}
\title{Portfolio Performance statistics}
\usage{
portfolio_stats(returns, market = NULL)
}
\arguments{
\item{returns}{xts. A time series object with returns for each asset in the portfolio.}

\item{market}{xts. A time series object with returns for the market index.}
}
\value{
A cleaned and processed data.table with price and return information.
}
\description{
This function calculates portfolio performance statistics such as cumulative return,
annualized return, Sharpe ratio, adjusted Sharpe ratio, maximum drawdown,
maximum drawdown length, and beta.
}
