% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_azure_blob.R
\name{save_azure_blob}
\alias{save_azure_blob}
\title{Save data to Azure Blob Storage}
\usage{
save_azure_blob(
  x,
  filename,
  container_name,
  endpoint = Sys.getenv("ENDPOINT"),
  key = Sys.getenv("KEY")
)
}
\arguments{
\item{x}{Data to save}

\item{filename}{Name of the file to save}

\item{container_name}{Name of the container to save the file to}

\item{endpoint}{Azure Blob Storage endpoint}

\item{key}{Azure Blob Storage key}
}
\value{
1
}
\description{
Save data to Azure Blob Storage
}
