% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wfo_combined}
\alias{wfo_combined}
\title{Combined Windowed Optimization with SMA Filtering}
\usage{
wfo_combined(df, params, window, window_type = "rolling")
}
\arguments{
\item{df}{DataFrame containing the data. Must include a `time` column and a `returns` column.}

\item{params}{DataFrame with parameters; expected to include columns "variable", "thresholds", and "sma_n".}

\item{window}{Integer specifying the window size (number of rows) to use for each optimization slice.}

\item{window_type}{A string to determine the window slicing method; either "rolling" (default) or "expanding".}
}
\value{
A list of DataFrames, each with a time stamp and the corresponding backtest results for that window.
}
\description{
Applies a window-based optimization to a DataFrame by slicing it into windows (either rolling or expanding)
and then performing SMA-based backtesting on each slice using [opt_with_sma_threshold]. The output is a list of
data frames binding the time stamp with the backtest results for each window.
}
